Move clientconfig/transformers to transformers/clientconfig
