+++ src/main/java/com/google/api/codegen/transformer/py/PythonClientConfigTransformer.java
@@ -1,0 +1,30 @@
+/* Copyright 2017 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.api.codegen.transformer.py;
+
+import com.google.api.codegen.gapic.GapicCodePathMapper;
+import com.google.api.codegen.transformer.clientconfig.AbstractClientConfigTransformer;
+
+/** The ModelToViewTransformer to transform a Model into the client config for Python. */
+public class PythonClientConfigTransformer extends AbstractClientConfigTransformer {
+  public PythonClientConfigTransformer(GapicCodePathMapper pathMapper) {
+    super(pathMapper);
+  }
+
+  @Override
+  protected String getExtension() {
+    return ".py";
+  }
+}
