+++ src/main/java/com/google/api/codegen/viewmodel/clientconfig/PairView.java
@@ -1,0 +1,38 @@
+/* Copyright 2017 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.api.codegen.viewmodel.clientconfig;
+
+import com.google.auto.value.AutoValue;
+
+/** Represents a key-value pair. */
+@AutoValue
+public abstract class PairView {
+  public abstract String key();
+
+  public abstract String value();
+
+  public static Builder newBuilder() {
+    return new AutoValue_PairView.Builder();
+  }
+
+  @AutoValue.Builder
+  public abstract static class Builder {
+    public abstract Builder key(String val);
+
+    public abstract Builder value(String val);
+
+    public abstract PairView build();
+  }
+}
