+++ src/main/java/com/google/api/codegen/viewmodel/clientconfig/RetryParamDefView.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.api.codegen.clientconfig.viewmodel;
+package com.google.api.codegen.viewmodel.clientconfig;
 
 import com.google.auto.value.AutoValue;
