--- src/main/resources/com/google/api/codegen/clientconfig/json.snip
@@ -1,13 +1,0 @@
-@extends "common.snip"
-@extends "clientconfig/common.snip"
-
-@snippet generateFilename(service)
-    {@context.upperCamelToLowerUnderscore(service.getSimpleName)}_client_config.json
-@end
-
-@snippet generateBody(service)
-    {
-      {@clientConfig(service)}
-    }
-
-@end
