--- src/main/resources/com/google/api/codegen/clientconfig/python_clientconfig.snip
@@ -1,13 +1,0 @@
-@extends "common.snip"
-@extends "clientconfig/common.snip"
-
-@snippet generateFilename(service)
-    {@context.upperCamelToLowerUnderscore(service.getSimpleName)}_client_config.py
-@end
-
-@snippet generateBody(service)
-    config = {
-      {@clientConfig(service)}
-    }
-
-@end
